Move expose api test to console basic test file
