+++ tests/ui-tests/tests/console-basic-test.js
@@ -1,13 +1,15 @@
+import request from 'request';
 import config from '../config';
 import kymaConsole from '../commands/console';
 import common from '../commands/common';
 import { describeIf } from '../utils/skip';
 import dex from '../utils/dex';
 import address from '../utils/address';
-import { retry } from '../utils/retry';
+import { retry, retryInterval } from '../utils/retry';
 import { testPluggable } from '../setup/test-pluggable';
+import { k8sApiNamespace, k8sApiDeployment, k8sApiService } from "./../setup/k8s-api"
 
-let page, browser;
+let page, browser, namespace;
 let token = '';
 
 // TODO: Move application tests to a separate file
@@ -23,7 +25,7 @@ describeIf(dex.isStaticUser(), 'Console basic tests', () => {
   });
 
   afterAll(async () => {
-    await kymaConsole.clearData(token, config.testNamespace);
+    await namespace.delete();
     if (browser) {
       await browser.close();
     }
@@ -136,4 +138,86 @@ describeIf(dex.isStaticUser(), 'Console basic tests', () => {
     }, 5);
     expect(applications).not.toContain(config.testApp);
   });
+
+  test('Expose API for Service', async () => {
+    const apiName = "ui-test-exposed-api";
+    let frame, exposedApiCellTexts, serviceUrl, service;
+
+    function callExposedAPI({ expectedStatusCode }) {
+      const req = {
+        url: `https://${apiName}.${config.domain}`,
+        method: 'GET',
+        // TODO: Analyze problem with UNABLE_TO_VERIFY_LEAF_SIGNATURE
+        rejectUnauthorized: false,
+      };
+
+      return new Promise((resolve, reject) => {
+        request(req, (error, response) => {
+          if (error) { reject(error) }
+          if (!response) { resolve("request returned no response") }
+          if (response.statusCode !== expectedStatusCode) {
+            resolve(`expected status code ${expectedStatusCode}, received ${response.statusCode}`)
+          }
+          resolve(true);
+        });
+      });
+    }
+
+    async function getCellsText() {
+      await frame.waitForSelector('[data-e2e-id=exposed-api-name]');
+      return {
+        name: (await kymaConsole.getNamesOnCurrentPage(page, "[data-e2e-id=exposed-api-name]"))[0],
+        secured: (await kymaConsole.getNamesOnCurrentPage(page, "[data-e2e-id=exposed-api-secured]"))[0],
+        idp: (await kymaConsole.getNamesOnCurrentPage(page, "[data-e2e-id=exposed-api-idp]"))[0]
+      }
+    }
+
+    // Create k8s resources
+    namespace = await new k8sApiNamespace();
+    await new k8sApiDeployment();
+    service = await new k8sApiService();
+
+    serviceUrl = address.console.getService(config.testNamespace, service.definition.metadata.name);
+
+    await Promise.all([
+      page.goto(serviceUrl),
+      page.waitForNavigation({
+        waitUntil: ['domcontentloaded', 'networkidle0'],
+      }),
+    ]);
+
+    // Before exposing API
+    console.log("Exposed API should retrieve 404 (does not exist)");
+    await callExposedAPI(404);
+
+    // Expose API (not secured)
+    frame = await kymaConsole.getFrame(page);
+    await frame.click("[data-e2e-id=open-expose-api]");
+    await frame.waitForSelector('#host-input');
+    await frame.type('#host-input', apiName);
+    await frame.click("[data-e2e-id=save-expose-api]");
+
+    exposedApiCellTexts = await getCellsText();
+    expect(exposedApiCellTexts.name).toEqual(` http-db-service-${apiName} `);
+    expect(exposedApiCellTexts.secured).toEqual(`No`);
+    expect(exposedApiCellTexts.idp).toEqual('-');
+
+    console.log("Exposed API should retrieve 200 (not secured)");
+    await retryInterval(callExposedAPI, { expectedStatusCode: 200 });
+
+    // Expose API (secured)
+    await kymaConsole.openLinkOnFrame(page, "[data-e2e-id=exposed-api-name]", `http-db-service-${apiName}`);
+    await frame.waitForSelector("[data-e2e-id=secure-api-checkbox]");
+    await frame.click("[data-e2e-id=secure-api-checkbox]");
+    await frame.click("[data-e2e-id=save-expose-api]");
+
+    exposedApiCellTexts = await getCellsText();
+    expect(exposedApiCellTexts.name).toEqual(` http-db-service-${apiName} `);
+    expect(exposedApiCellTexts.secured).toEqual(`Yes`);
+    expect(exposedApiCellTexts.idp).toEqual('DEX');
+
+    console.log("Exposed API should retrieve 401 (secured)");
+    await retryInterval(callExposedAPI, { expectedStatusCode: 401 });
+  });
+
 });
